backend/requirements.txt
# Core Django
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
django-extensions==3.2.3
django-celery-beat==2.6.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cache / Queue
redis==5.0.3
celery==5.3.6

# Storage
boto3==1.34.69
django-storages==1.14.2

# PDF Generation
reportlab==4.1.0
Pillow==10.2.0

# HTTP
requests==2.31.0
httpx==0.27.0

# AI
anthropic==0.23.1

# Utilities
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0

# Dev / Testing
pytest-django==4.8.0
factory-boy==3.3.0